Test fixes, all passing.
Last commit I jumped the gun on references to File.split that
weren't working, thinking it was due to a Ruby upgrade and/or dependency
on `ftools`. In fact it was due to namespace collision between `ClWiki::File`
and the stdlib `File`. I put things back and the tests are passing again.

I do need to deal with the File name collision - it's too ridiculous
having to have `::File` references everywhere.