All tests passing again.
While doing @routes = ClWiki::Engine.routes in a before block 'twould
still appear to be legit (and less noisy), decided to go with the
documented approach of passing use_route: in each request.

The redirect tests went wonky when on a recent commit I started using
foo_url instead of foo_path, because I had a passing test within the
legacy tests that didn't notice I wasn't redirecting in the prod code
with foo_url. Switching from foo_path to foo_url in the asserts before I
converted the whole project from an App to an Engine was seamless ...
but then broke afterwards, presumably because of some missing foo in the
test setup bowels? Dunno. But switching back to foo_path in the asserts
works without having to check the host ... maybe that's a mistake in the
legacy tests, but ... it's late.

I did find a real problem, though, now that I've switched to an Engine
and am running the tests through a dummy application which has the
engine mounted on a parent path: the legacy check was looking at the
request.path_info, which didn't include the parent path portion and no
longer worked. Now that's fixed.