Added metadata handling of mtime.
It's a sorta-support-any-kind-of-metadata-code mixed with mtime
specifics that need to affect things being read and written and all -
half wondering if it should have been mixed in, just to help keep things
a little cleaner ... not sure. But it appears to be working fine. The
goal being the metadata mod time overrides the actual file's mtime
value. This allows storage of the mtime to be accurate to when the wiki
originally writes the file, and then the files be checked in/out,
exported, moved, copied, losing the OS mtime, but keeping the original
date of the authoring of the page.