recent page name link change && bundle update
The recent page view used to use the default page header for each page,
which made the page name link to the find page, to find any references
to that page. But from the recent page view, that doesn't make much
sense. Now the link is to the whole page view itself.