Important bug fix in legacy file parsing.
This could just be a small bug introduced in 2.1.6 of clWiki when
storing page content metadata incorrectly. Pages in 2.1.7 were not at
all parsing the metadata correctly and flooding the file itself with
duplicates of the metadata on save - which was at least visible, and
didn't lose any data, but flooded it with copies of the entire page as
if it was also metadata.