Refactor parsing metadata from file.
I tried a different implementation, the current one looking
overwrought, but after trying something new, this is fairly succinct
seeming for what it's doing. I'm sure there's probably a tighter way to
do it, but it's not as bad as I thought.

The File class is a little cleaner now, having re-arranged things a
little.