The big index reckoning! [Breaking - 3.0.0]
No more crazy indexing options. This is a breaking change, ripping out
all of these index options that I don't want anymore now that I've
decided to go whole hog on the new Memory Index. This limited the need
to support @owner for accessing pages to just a few places.

This is also a half-hearted attempt to simplify many of the options that
are also no longer needed in configuration.rb ... there's probably a lot
more that can happen there.

There's one odd edge-case left in here. Since we're using the index for
determining `page_exists?` there's a call stack when adding a new page
before the index has been built, where we don't have the proper
ownership to establish. There may be a simple re-design here, but
basically it's through a class method on Page, so no @owner instance
exists there statically.

And there doesn't _need_ to be one if the Index was flexible enough to
lazy-build itself - page existence doesn't require a User for accessing
possibly encrypted contents.

But - that's (probably) not a problem I'll need to solve in the real
life instances (just the 2 I know of for myself personally) of this wiki
engine. So - the bug still exists, but the test tweaks the setup to make
sure the index has already been built and avoid the bug.