Fix title_only search. Requires fixes.
Running lib tests in isolation didn't work with last round of changes to
clean up circular requires.

The bug worked-around that probably won't happen, actually is the first
thing to happen, because we use the index for page_exists? in parsing
the home page first loaded after first boot.

Added a TODO to configuration.rb, I should just embrace a single-owner
wiki because that's all I need, and that should solve this problem at a
better place ... maybe.

Need to play out the use cases for authentication, if it must match the
only owner configured. Probably simple enough.