UI for toggling encryption on/off on page edit
This includes a bug fix File that ensures the contents are returned now
always as a string. It used to return as an Array, then an encrypted
file returns its contents as a String, but until this got hooked into
the UI, the inconsistency hadn't been revealed in tests.

This is another breaking change for 3.0, having a String returned
instead of Array for the legacy (non-encrypted) case. But it makes
better sense because there were many call sites that had to handle the
transformation from Array to String.

There's a new configuration option as well to determine if encrypting is
the default option for new pages.