Bug fix to ensure utf-8 file contents.
I couldn't easily get a failing test for this, but I had a wiki with a
custom formatter for smart quotes that couldn't work because of the
change in 3.0 to using `File.binread` for encryption cases. That leaves
the string in an ASCII-8BIT encoding, which isn't really an encoding,
but just an indicator that the bytes are binary.

Now, when just working with unencrypted data, the content is properly
retrieved in UTF-8 (the default encoding with `File.read`).